app/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
jinja2==3.1.3
python-dotenv==1.0.1
langchain==0.1.9
langgraph==0.0.27
langchain-groq==0.0.6
langchain-community==0.0.24
numpy==1.26.4
pandas==2.2.0
joblib==1.3.2
scikit-learn==1.4.0
tavily-python==0.3.1
tensorflow==2.15.0
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
scipy==1.11.4
h5py==3.10.0
pillow==10.2.0
requests==2.31.0
tqdm==4.66.1
